# openapi.yaml
openapi: 3.0.3
info:
  title: School Reading & Assessment Platform API
  version: 1.0.0
  description: |
    Complete REST API for the school oral reading and AI assessment platform.

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a JWT Bearer token
    in the `Authorization: Bearer <token>` header. Access tokens expire in 15 minutes.
    Use `/auth/refresh` to obtain a new one without re-logging in.

    ## Roles
    | Role    | Access                                                        |
    |---------|---------------------------------------------------------------|
    | admin   | School structure, user management, all assignments            |
    | teacher | Their assigned sections, lesson upload, student progress view |
    | student | Their section's lessons, reading sessions, personal dashboard |

    ## Async Assessment Flow
    1. `POST /student/sessions` — creates session record
    2. Student records audio in browser via MediaRecorder API
    3. `POST /student/sessions/{id}/submit` — uploads audio (WebM/OGG)
    4. Server streams audio to Whisper STT microservice (audio never stored)
    5. Transcript sent to Ollama LLM for assessment (BullMQ job)
    6. **Socket.io** emits `assessment.complete` to student's browser
    7. `GET /student/sessions/{id}/result` — fetch the assessment result

    ## Socket.io Events
    Connect with the JWT token as a query param: `ws://server/api/v1?token=<jwt>`

...
servers:
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login

  # ── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    SectionIdParam:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StudentIdParam:
      name: studentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-05-30T08:30:00Z"
...
    UserRole:
      type: string
      enum: [admin, teacher, student]
      example: student
...
    Pagination:
      type: object
      required: [total, page, limit, total_pages]
      properties:
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          nullable: true
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: must be a valid email address
    ValidationError:
      type: object
      required: [code, message, errors]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    # ── Auth Schemas ──────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [full_name, email, password, role]
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 255
          example: Priya Sharma
        email:
          type: string
          format: email
          example: priya@school.edu
        password:
          type: string
          minLength: 8
          maxLength: 72
          example: SecurePass@123
          description: Minimum 8 characters; stored as bcrypt hash
        role:
          type: string
          enum: [student, teacher]
          example: student
          description: Choose your account type. Admin accounts are provisioned separately.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@school.edu
        password:
          type: string
          example: SecurePass@123
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...

    # ── User Schemas ──────────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, full_name, email, role, is_active, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        full_name:
          type: string
          example: Priya Sharma
        email:
          type: string
          example: priya@school.edu
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          example: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/AuthTokens'
        - type: object
          required: [user]
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
    AdminUserView:
      type: object
      required: [id, full_name, email, role, is_active, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        full_name:
          type: string
          example: Ravi Kumar
        email:
          type: string
          example: ravi@school.edu
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          example: true
        assigned_section:
          description: Populated only when role = student
          nullable: true
          type: object
          properties:
            section_id:
              $ref: '#/components/schemas/UUID'
            section_name:
              type: string
              example: Section B
            class_name:
              type: string
              example: Grade 4
        assigned_sections:
          description: Populated only when role = teacher
          nullable: true
          type: array
          items:
            type: object
            properties:
              section_id:
                $ref: '#/components/schemas/UUID'
              section_name:
                type: string
                example: Section A
              class_name:
                type: string
                example: Grade 5
        created_at:
          $ref: '#/components/schemas/Timestamp'

    # ── Class Schemas ─────────────────────────────────────────────────────
...
    SkillTrend:
      type: object
      required: [date, accuracy_score, fluency_score, pronunciation_score, overall_score]
      properties:
        date:
          type: string
          format: date
          example: "2026-05-30"
        accuracy_score:
          type: number
          format: float
          example: 87.5
        fluency_score:
          type: number
          format: float
          example: 74.0
        pronunciation_score:
          type: number
          format: float
          example: 69.5
        overall_score:
          type: number
          format: float
          example: 78.2
...
    StudentDashboard:
      type: object
      required: [student, streak, progress, skill_profile, recent_sessions]
      properties:
        student:
          type: object
          required: [id, full_name, section_name, class_name]
          properties:
            id:
              $ref: '#/components/schemas/UUID'
            full_name:
              type: string
              example: Arun Prakash
            section_name:
              type: string
              example: Section A
            class_name:
              type: string
              example: Grade 5
        streak:
          $ref: '#/components/schemas/StreakInfo'
        current_rank:
          type: integer
          nullable: true
          example: 3
          description: Current week leaderboard rank within section
        progress:
          type: object
          required: [overall_progress_percent, lessons_completed, total_lessons, total_sessions]
          properties:
            overall_progress_percent:
              type: number
              format: float
              example: 60.0
              description: "(lessons with ≥1 completed session / total published lessons) × 100"
            lessons_completed:
              type: integer
              example: 6
            total_lessons:
              type: integer
              example: 10
            total_sessions:
              type: integer
              example: 14
        skill_profile:
          $ref: '#/components/schemas/StudentSkillProfile'
        recent_sessions:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/RecentSession'
        next_lesson:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/NextLesson'
...
    StudentProgressDetail:
      type: object
      required: [student, skill_profile, skill_trend, session_history]
      properties:
        student:
          type: object
          required: [id, full_name, email]
          properties:
            id:
              $ref: '#/components/schemas/UUID'
            full_name:
              type: string
              example: Arun Prakash
            email:
              type: string
              example: arun@school.edu
        skill_profile:
          $ref: '#/components/schemas/StudentSkillProfile'
        skill_trend:
          type: array
          description: One data point per session, up to the requested number of days
          items:
            $ref: '#/components/schemas/SkillTrend'
        session_history:
          type: array
          items:
            type: object
            required: [session_id, lesson_title, attempt_number, completed_at]
            properties:
              session_id:
                $ref: '#/components/schemas/UUID'
              lesson_id:
                $ref: '#/components/schemas/UUID'
              lesson_title:
                type: string
                example: Chapter 3 - The Water Cycle
              attempt_number:
                type: integer
                example: 2
              overall_score:
                type: number
                format: float
                nullable: true
                example: 78.2
              accuracy_score:
                type: number
                format: float
                nullable: true
                example: 84.5
              fluency_score:
                type: number
                format: float
                nullable: true
                example: 71.0
              pronunciation_score:
                type: number
                format: float
                nullable: true
                example: 68.5
              duration_seconds:
                type: integer
                nullable: true
                example: 187
              completed_at:
                $ref: '#/components/schemas/Timestamp'
...
    StudentWeeklyLeaderboard:
      allOf:
        - $ref: '#/components/schemas/WeeklyLeaderboard'
        - type: object
          properties:
            my_rank:
              type: integer
              nullable: true
              example: 7
              description: Current user's rank; null if no sessions this week

  # ── Reusable Responses ────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired token
    Forbidden:
      description: Authenticated but insufficient role permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Resource not found
...
    UnprocessableEntity:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      description: |
        Creates an account with the specified role (student or teacher). The account
        is inactive until an admin approves it. An admin must also assign a section
        before the user can access role-specific features.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              full_name: Priya Sharma
              email: priya@school.edu
              password: SecurePass@123
              role: student
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                required: [message, user_id]
                properties:
                  message:
                    type: string
                    example: Account created. Awaiting admin approval before you can log in.
                  user_id:
                    $ref: '#/components/schemas/UUID'
        '409':
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: EMAIL_TAKEN
                message: An account with this email already exists
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT access + refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@school.edu
              password: SecurePass@123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or inactive account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_credentials:
                  summary: Wrong email or password
                  value:
                    code: INVALID_CREDENTIALS
                    message: Email or password is incorrect
                inactive:
                  summary: Account pending approval or deactivated
                  value:
                    code: ACCOUNT_INACTIVE
                    message: Account pending admin approval. Contact your admin.
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          description: Too many failed login attempts (5 per 15 minutes per IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: RATE_LIMITED
                message: Too many login attempts. Try again in 15 minutes.
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                required: [access_token, expires_in]
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  expires_in:
                    type: integer
                    example: 900
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke the refresh token
      description: Access token expires naturally after 15 minutes; the refresh token is immediately invalidated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user's profile
      responses:
        '200':
          description: Authenticated user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ═══════════════════════════════════════════════════════════════════════════
  # ADMIN — CLASSES
  # ═══════════════════════════════════════════════════════════════════════════
...
  /admin/users:
    get:
      tags: [Admin — Users]
      summary: List all registered users
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: role
          in: query
          description: Filter by role
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: unassigned
          in: query
          description: If true, return only users without a section assignment
          schema:
            type: boolean
        - name: search
          in: query
          description: Partial match on name or email
          schema:
            type: string
            example: priya
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserView'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}:
    get:
      tags: [Admin — Users]
      summary: Get a single user's full profile
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/role:
    patch:
      tags: [Admin — Users]
      summary: Assign or change a user's role
      description: Cannot be used to assign or change the admin role.
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [teacher, student]
                  example: student
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserView'
        '400':
          description: Cannot modify admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_OPERATION
                message: Admin role cannot be changed via this endpoint
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/status:
    patch:
      tags: [Admin — Users]
      summary: Activate or deactivate a user account
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserView'
        '400':
          description: Cannot deactivate own admin account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_OPERATION
                message: You cannot deactivate your own account
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # ADMIN — TEACHER ASSIGNMENTS
  # ═══════════════════════════════════════════════════════════════════════════
...
  /teacher/sections/{sectionId}/students/{studentId}/progress:
    get:
      tags: [Teacher — Dashboard]
      summary: Drilldown — full progress detail for one student
      parameters:
        - $ref: '#/components/parameters/SectionIdParam'
        - $ref: '#/components/parameters/StudentIdParam'
        - name: days
          in: query
          description: Days of history for skill trend chart (7-90)
          schema:
            type: integer
            minimum: 7
            maximum: 90
            default: 30
      responses:
        '200':
          description: Student progress detail with session history and skill trend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProgressDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Section or student not found, or student not in this section
          $ref: '#/components/responses/NotFound'
...
  /student/dashboard:
    get:
      tags: [Student — Dashboard]
      summary: Get the student's full dashboard
      description: |
        Returns streak, current rank, overall progress, skill profile,
        5 most recent sessions, and the next recommended lesson.
        All data is scoped to the student's active section.
      responses:
        '200':
          description: Student dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          description: Student has not been assigned to a section yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_ASSIGNED
                message: You have not been assigned to a section yet.
  /student/leaderboard:
    get:
      tags: [Student — Dashboard]
      summary: Get the weekly leaderboard for the student's section
      description: |
        Returns all entries for the requested ISO week.
        `is_current_user: true` marks the authenticated student's row.
        `my_rank` is also returned at the top level for students outside the top 10.
        Leaderboard resets each Monday at 00:00 server time.
      parameters:
        - name: week_start
          in: query
          description: "ISO Monday date (default: current week's Monday)"
          schema:
            type: string
            format: date
            example: "2026-05-25"
      responses:
        '200':
          description: Weekly leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentWeeklyLeaderboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          description: Student has not been assigned to a section yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_ASSIGNED
                message: You have not been assigned to a section yet.
  /student/progress:
    get:
      tags: [Student — Dashboard]
      summary: Get skill trend data points for dashboard charts
      description: |
        Returns one data point per completed session within the requested time window.
        Use this for the line charts on the student dashboard (accuracy, fluency, pronunciation over time).
      parameters:
        - name: days
          in: query
          description: Number of past days to include (7-90)
          schema:
            type: integer
            minimum: 7
            maximum: 90
            default: 30
      responses:
        '200':
          description: Skill trend data and improvement summary
          content:
            application/json:
              schema:
                type: object
                required: [trend, summary]
                properties:
                  trend:
                    type: array
                    items:
                      $ref: '#/components/schemas/SkillTrend'
                  summary:
                    type: object
                    required: [best_overall_score]
                    properties:
                      best_overall_score:
                        type: number
                        format: float
                        nullable: true
                        example: 91.0
                      most_improved_skill:
                        type: string
                        nullable: true
                        example: fluency
                        description: Skill with the largest positive delta over the period
                      most_improved_delta:
                        type: number
                        format: float
                        nullable: true
                        example: 12.5
                        description: Score improvement (last session minus first session in window)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          description: Student has not been assigned to a section yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: NOT_ASSIGNED
                message: You have not been assigned to a section yet.